% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{citron_palettes}
\alias{citron_palettes}
\title{Thesis colour palettes.
This code is based pretty heavily on the 'ochRe' package.
Colours are inspired by Jacquemus' Spring 2019 RTW collection, 'La Riviera'.}
\format{An object of class \code{list} of length 5.}
\usage{
citron_palettes
}
\description{
Thesis colour palettes.
This code is based pretty heavily on the 'ochRe' package.
Colours are inspired by Jacquemus' Spring 2019 RTW collection, 'La Riviera'.
}
\keyword{datasets}
